Refactor ssot to robot parser
