+++ server/services/SsotToRobotParsing/robotCodeConstants.js
@@ -1,0 +1,13 @@
+const ACTIVITY_IDENTIFIER = 'INSTRUCTION';
+const LINEBREAK = '\n';
+const COMMENT = '#';
+const DOUBLESPACE = '  ';
+const FOURSPACE = '    ';
+
+module.exports = {
+  ACTIVITY_IDENTIFIER,
+  LINEBREAK,
+  COMMENT,
+  DOUBLESPACE,
+  FOURSPACE,
+};
